Changes: 1. Segregate the needed function to the corresponding .c and .h files.
