+++ ListSLL.h
@@ -8,5 +8,6 @@ struct node{
 
 void create_node(int _x);
 void insert(int _x,struct node *_L);
+void browse(void);
 
 #endif
